+++ CHANGELOG.md
@@ -9,7 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-* None yet!
+* Update keycloak image to quay
+* Update auth examples for new keycloak handling
+* Upgrade pulumi and keycloak setup
 
 ## [0.27.2] - 10/06/2024